Start rework from template to Youtube2Mp3 app Add BrowseSongsPage, DownloadedSongsPage and SpotifyPage, refactor a few files, fix naming, etc ...
