+++ src/Youtube2Mp3.Xamarin/Youtube2Mp3.Xamarin/Views/DownloadedSongsPage.xaml
@@ -7,9 +7,6 @@
              x:Class="Youtube2Mp3.Xamarin.Views.DownloadedSongsPage">
     <ContentPage.Content>
         <StackLayout>
-            <Label Text="Welcome to Xamarin.Forms!"
-                VerticalOptions="CenterAndExpand" 
-                HorizontalOptions="CenterAndExpand" />
         </StackLayout>
     </ContentPage.Content>
 </ContentPage>
