--- src/Youtube2Mp3.Xamarin/Youtube2Mp3.Xamarin/Views/ItemsPage.xaml.cs
@@ -1,48 +1,0 @@
-﻿using System;
-using System.ComponentModel;
-using Xamarin.Forms;
-using Youtube2Mp3.Xamarin.Models;
-using Youtube2Mp3.Xamarin.ViewModels;
-
-namespace Youtube2Mp3.Xamarin.Views
-{
-    // Learn more about making custom code visible in the Xamarin.Forms previewer
-    // by visiting https://aka.ms/xamarinforms-previewer
-    [DesignTimeVisible(false)]
-    public partial class ItemsPage : ContentPage
-    {
-        ItemsViewModel viewModel;
-
-        public ItemsPage()
-        {
-            InitializeComponent();
-
-            BindingContext = viewModel = new ItemsViewModel();
-        }
-
-        async void OnItemSelected(object sender, SelectedItemChangedEventArgs args)
-        {
-            var item = args.SelectedItem as Item;
-            if (item == null)
-                return;
-
-            await Navigation.PushAsync(new ItemDetailPage(new ItemDetailViewModel(item)));
-
-            // Manually deselect item.
-            ItemsListView.SelectedItem = null;
-        }
-
-        async void AddItem_Clicked(object sender, EventArgs e)
-        {
-            await Navigation.PushModalAsync(new NavigationPage(new NewItemPage()));
-        }
-
-        protected override void OnAppearing()
-        {
-            base.OnAppearing();
-
-            if (viewModel.Items.Count == 0)
-                viewModel.LoadItemsCommand.Execute(null);
-        }
-    }
-}
