+++ src/Youtube2Mp3.Xamarin/Youtube2Mp3.Xamarin/Views/SearchSongsPage.xaml
@@ -4,34 +4,17 @@
              xmlns:d="http://xamarin.com/schemas/2014/forms/design"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d"
-             x:Class="Youtube2Mp3.Xamarin.Views.ItemsPage"
-             Title="{Binding Title}"
-             x:Name="BrowseItemsPage">
-
-    <ContentPage.ToolbarItems>
-        <ToolbarItem Text="Add" Clicked="AddItem_Clicked" />
-    </ContentPage.ToolbarItems>
-
+             x:Class="Youtube2Mp3.Xamarin.Views.SearchSongsPage"
+             x:Name="BrowseSongsPage">
+    
     <StackLayout>
-        <ListView x:Name="ItemsListView"
+        <Entry Placeholder="Search query" />
+        <ListView x:Name="SearchResults"
                 ItemsSource="{Binding Items}"
                 VerticalOptions="FillAndExpand"
                 HasUnevenRows="true"
-                RefreshCommand="{Binding LoadItemsCommand}"
-                IsPullToRefreshEnabled="true"
-                IsRefreshing="{Binding IsBusy, Mode=OneWay}"
                 CachingStrategy="RecycleElement"
                 ItemSelected="OnItemSelected">
-            <d:ListView.ItemsSource>
-                <x:Array Type="{x:Type x:String}">
-                    <x:String>First Item</x:String>
-                    <x:String>Second Item</x:String>
-                    <x:String>Third Item</x:String>
-                    <x:String>Forth Item</x:String>
-                    <x:String>Fifth Item</x:String>
-                    <x:String>Sixth Item</x:String>
-                </x:Array>
-            </d:ListView.ItemsSource>
             <ListView.ItemTemplate>
                 <DataTemplate>
                     <ViewCell>
@@ -40,7 +23,7 @@
                                 d:Text="{Binding .}"
                                 LineBreakMode="NoWrap" 
                                 Style="{DynamicResource ListItemTextStyle}" 
-                                FontSize="16" />
+                                FontSize="17" />
                             <Label Text="{Binding Description}" 
                                 d:Text="Item description"
                                 LineBreakMode="NoWrap"
@@ -52,5 +35,5 @@
             </ListView.ItemTemplate>
         </ListView>
     </StackLayout>
-
+    
 </ContentPage>
